move DashboardMetadataProvider to org.wso2.carbon.dashboards.core
